Bugfix for send bytes api delegate

+++ TestApps/iPhoneTestApp/Classes/iPhoneTestAppViewController.m
@@ -140,10 +140,10 @@
 -(void)soundCloudAPI:(SCSoundCloudAPI *)api didReceiveData:(NSData *)data context:(id)context;
 {}
 
--(void)soundCloudAPI:(SCSoundCloudAPI *)api didReceiveBytes:(UInt32)loadedBytes total:(UInt32)totalBytes context:(id)context;
+-(void)soundCloudAPI:(SCSoundCloudAPI *)api didReceiveBytes:(unsigned long long)loadedBytes total:(unsigned long long)totalBytes context:(id)context;
 {}
 
--(void)soundCloudAPI:(SCSoundCloudAPI *)api didSendBytes:(UInt32)sendBytes total:(UInt32)totalBytes context:(id)context;
+-(void)soundCloudAPI:(SCSoundCloudAPI *)api didSendBytes:(unsigned long long)sendBytes total:(unsigned long long)totalBytes context:(id)context;
 {
 	if([context isEqual:@"upload"]) {
 		[progresBar setProgress: ((float)sendBytes) / totalBytes];
